+++ all/atob/index.browser.js
@@ -1,1 +1,12 @@
-export * from './index.js'
+/**
+ * @param {string}str
+ * @returns {string}
+ */
+export const atob = globalThis.atob
+/**
+ * @return {typeof atob}
+ */
+export function atob_() {
+	return atob
+}
+export { atob_ as _atob, }