+++ web_animation/index.browser.d.ts
@@ -1,0 +1,1 @@
+export * from '../all/wanimato/index.js'